+++ src/symphony-core/NIDAQController/NIDAQOutputStream.cs
@@ -114,11 +114,11 @@ namespace NI
                         throw new DAQException(ed.Name + " background must contain a value of 0 or 1");
 
                     ushort bitPosition = BitPositions[ed];
-                    m = MeasurementPool.GetMeasurement((int)((int)m.QuantityInBaseUnits << bitPosition), 0, m.BaseUnits);
+                    m = MeasurementPool.GetMeasurement((long)((long)m.QuantityInBaseUnits << bitPosition), 0, m.BaseUnits);
 
                     background = background == null
                         ? m
-                        : MeasurementPool.GetMeasurement((int)background.QuantityInBaseUnits | (int)m.QuantityInBaseUnits, 0, background.BaseUnits);
+                        : MeasurementPool.GetMeasurement((long)background.QuantityInBaseUnits | (long)m.QuantityInBaseUnits, 0, background.BaseUnits);
                 }
                 return background;
             }
@@ -140,14 +140,14 @@ namespace NI
                     if (m.QuantityInBaseUnits != 0 && m.QuantityInBaseUnits != 1)
                         throw new DAQException(ed.Name + " output data must contain only values of 0 and 1");
 
-                    return MeasurementPool.GetMeasurement((int)((int)m.QuantityInBaseUnits << bitPosition), 0, m.BaseUnits);
+                    return MeasurementPool.GetMeasurement((long)((long)m.QuantityInBaseUnits << bitPosition), 0, m.BaseUnits);
                 }));
 
                 outData = outData == null
                               ? pulled
                               : outData.Zip(pulled,
                                             (m1, m2) =>
-                                            MeasurementPool.GetMeasurement((int)m1.QuantityInBaseUnits | (int)m2.QuantityInBaseUnits, 0, m1.BaseUnits));
+                                            MeasurementPool.GetMeasurement((long)m1.QuantityInBaseUnits | (long)m2.QuantityInBaseUnits, 0, m1.BaseUnits));
             }
 
             if (!outData.SampleRate.Equals(this.SampleRate))
